# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: 电商平台 API
  description: |
    3 人小组项目共用接口契约。修改接口请先改本文档，再实现前后端。
    业务术语见 docs/BUSINESS_GLOSSARY.md，状态机见 docs/domain/DOMAIN_MODEL.md
    管理后台见 docs/domain/ADMIN.md
  version: 0.2.0
servers:
  - url: http://localhost:8080/api
    description: 本地开发（端口按实际后端调整）
...
paths:
...
  /products:
    get:
      tags: [product]
      summary: 商品列表（SPU）
      parameters:
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductListResponse'
...
  /users/addresses:
    get:
      tags: [user]
      summary: 当前用户收货地址列表
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Address'
    post:
      tags: [user]
      summary: 新增收货地址
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAddressRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Address'
  /users/addresses/{addressId}:
    put:
      tags: [user]
      summary: 修改收货地址
      security:
        - bearerAuth: []
      parameters:
        - name: addressId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAddressRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Address'
    delete:
      tags: [user]
      summary: 删除收货地址
      security:
        - bearerAuth: []
      parameters:
        - name: addressId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: No Content
  /users/addresses/{addressId}/default:
    patch:
      tags: [user]
      summary: 设为默认地址
      security:
        - bearerAuth: []
      parameters:
        - name: addressId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Address'

  # ========== Order ==========
  /orders:
    post:
      tags: [order]
      summary: 创建订单（从购物车或直接购买）
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: 订单创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
    get:
      tags: [order]
      summary: 用户订单列表
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/OrderStatus'
      responses:
        '200':
          description: OK
  /orders/{orderId}:
    get:
      tags: [order]
      summary: 订单详情
      security:
        - bearerAuth: []
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
  /orders/{orderId}/pay:
    post:
      tags: [order]
      summary: Mock 支付
      description: 模拟支付成功，订单状态 PENDING_PAYMENT → PAID → PENDING_SHIPMENT
      security:
        - bearerAuth: []
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 支付成功
  /orders/{orderId}/cancel:
    post:
      tags: [order]
      summary: 取消订单（仅待支付）
      security:
        - bearerAuth: []
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
...
  /merchant/applications:
    post:
      tags: [merchant]
      summary: 提交商家入驻申请（无需登录）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MerchantApplicationSubmitRequest'
      responses:
        '201':
          description: Created
        '409':
          description: 重复申请或已入驻
  /merchant/applications/status:
    get:
      tags: [merchant]
      summary: 按手机号查询入驻申请进度（无需登录）
      parameters:
        - name: contactPhone
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
  /merchant/dashboard/summary:
    get:
      tags: [merchant]
      summary: 商家工作台统计
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MerchantDashboardSummary'
  /merchant/products:
    get:
      tags: [merchant]
      summary: 商家商品列表
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MerchantProductListResponse'
    post:
      tags: [merchant]
      summary: 创建商品（SPU+SKU）
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMerchantProductRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MerchantProduct'
  /merchant/products/{spuId}/submit-audit:
    post:
      tags: [merchant]
      summary: 提交商品审核
      description: 商品状态 DRAFT → PENDING_AUDIT
      security:
        - bearerAuth: []
      parameters:
        - name: spuId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MerchantProductSubmitAuditResponse'
  /merchant/orders:
    get:
      tags: [merchant]
      summary: 商家订单列表
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/OrderStatus'
      responses:
        '200':
          description: OK
  /merchant/orders/{subOrderId}/ship:
    post:
      tags: [merchant]
      summary: 发货
      security:
        - bearerAuth: []
      parameters:
        - name: subOrderId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [logisticsCompany, trackingNo]
              properties:
                logisticsCompany:
                  type: string
                trackingNo:
                  type: string
      responses:
        '200':
          description: OK
  /merchant/after-sales:
    get:
      tags: [merchant]
      summary: 商家售后列表
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/AfterSaleStatus'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AfterSaleListResponse'
  /merchant/after-sales/{afterSaleId}/audit:
    post:
      tags: [merchant]
      summary: 商家审核售后
      security:
        - bearerAuth: []
      parameters:
        - name: afterSaleId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AfterSaleAuditRequest'
      responses:
        '200':
          description: OK

  # ========== Admin ==========
...
  /admin/orders:
    get:
      tags: [admin]
      summary: 全平台订单列表
      security:
        - bearerAuth: []
      parameters:
        - name: orderNo
          in: query
          schema:
            type: string
        - name: userId
          in: query
          schema:
            type: integer
        - name: merchantId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/OrderStatus'
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderListResponse'
  /admin/orders/{orderId}:
    get:
      tags: [admin]
      summary: 订单详情
      security:
        - bearerAuth: []
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDetail'
...
  /admin/merchants/applications:
    get:
      tags: [admin]
      summary: 商家入驻申请列表（可按状态筛选）
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            description: PENDING / APPROVED / REJECTED / ALL，多个用逗号分隔
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MerchantApplicationListResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    AfterSaleType:
      type: string
      enum:
        - REFUND_ONLY
        - RETURN_REFUND
    AfterSaleStatus:
      type: string
      enum:
        - APPLIED
        - APPROVED
        - REJECTED
        - ESCALATED
        - RETURNING
        - REFUNDED
...
    AfterSaleAuditRequest:
      type: object
      required: [approved]
      properties:
        approved:
          type: boolean
        reason:
          type: string
    AfterSale:
      type: object
      properties:
        afterSaleId:
          type: integer
        orderId:
          type: integer
        orderNo:
          type: string
        subOrderId:
          type: integer
        merchantId:
          type: integer
        userId:
          type: integer
        shopName:
          type: string
        type:
          $ref: '#/components/schemas/AfterSaleType'
        reason:
          type: string
        status:
          $ref: '#/components/schemas/AfterSaleStatus'
          description: ESCALATED 表示待平台仲裁
        appliedAt:
          type: string
          format: date-time
        merchantDeadline:
          type: string
          format: date-time
        auditReason:
          type: string
        items:
          type: array
          items:
            type: object
            properties:
              skuId:
                type: integer
              title:
                type: string
              price:
                type: number
              quantity:
                type: integer
    AfterSaleListResponse:
      type: object
      properties:
        total:
          type: integer
        list:
          type: array
          items:
            $ref: '#/components/schemas/AfterSale'
...
    PendingMerchant:
      type: object
      properties:
        merchantId:
          type: integer
          description: 申请 ID
        shopName:
          type: string
        contactName:
          type: string
        contactPhone:
          type: string
        appliedAt:
          type: string
          format: date-time
    MerchantApplicationSubmitRequest:
      type: object
      required: [shopName, contactName, contactPhone]
      properties:
        shopName:
          type: string
        contactName:
          type: string
        contactPhone:
          type: string
    MerchantApplication:
      allOf:
        - $ref: '#/components/schemas/PendingMerchant'
        - type: object
          properties:
            status:
              type: string
              enum: [PENDING, APPROVED, REJECTED]
            auditedAt:
              type: string
              format: date-time
            rejectReason:
              type: string
            approvedMerchantId:
              type: integer
            merchantUsername:
              type: string
              description: 审核通过后生成的商家登录账号
    MerchantApplicationListResponse:
      type: object
      properties:
        total:
          type: integer
        list:
          type: array
          items:
            $ref: '#/components/schemas/MerchantApplication'
    OrderListResponse:
      type: object
      properties:
        total:
          type: integer
        list:
          type: array
          items:
            $ref: '#/components/schemas/Order'
    OrderDetail:
      allOf:
        - $ref: '#/components/schemas/Order'
        - type: object
          properties:
            subOrders:
              type: array
              items:
                $ref: '#/components/schemas/SubOrder'
            addressSnapshot:
              type: object
              properties:
                receiverName:
                  type: string
                phone:
                  type: string
                fullAddress:
                  type: string
    SubOrder:
      type: object
      properties:
        subOrderId:
          type: integer
        merchantId:
          type: integer
        shopName:
          type: string
        status:
          $ref: '#/components/schemas/OrderStatus'
        shipment:
          type: object
          properties:
            logisticsCompany:
              type: string
            trackingNo:
              type: string
            shippedAt:
              type: string
              format: date-time
    ProductStatus:
      type: string
      enum:
        - DRAFT
        - PENDING_AUDIT
        - ON_SHELF
        - REJECTED
        - OFF_SHELF
    MerchantDashboardSummary:
      type: object
      properties:
        merchantId:
          type: integer
        shopId:
          type: integer
        shopName:
          type: string
        productTotal:
          type: integer
        draftProductCount:
          type: integer
        pendingAuditProductCount:
          type: integer
        onShelfProductCount:
          type: integer
        rejectedProductCount:
          type: integer
        offShelfProductCount:
          type: integer
        pendingShipmentOrderCount:
          type: integer
        shippedOrderCount:
          type: integer
    MerchantProductListResponse:
      type: object
      properties:
        total:
          type: integer
        list:
          type: array
          items:
            $ref: '#/components/schemas/MerchantProduct'
    MerchantProduct:
      type: object
      properties:
        spuId:
          type: integer
        shopId:
          type: integer
        shopName:
          type: string
        merchantId:
          type: integer
        categoryId:
          type: integer
        categoryName:
          type: string
          nullable: true
        title:
          type: string
        description:
          type: string
        mainImage:
          type: string
        status:
          $ref: '#/components/schemas/ProductStatus'
        submittedAt:
          type: string
          format: date-time
        rejectReason:
          type: string
          description: 平台驳回原因；仅 status=REJECTED 时有值
        skus:
          type: array
          items:
            $ref: '#/components/schemas/MerchantSku'
    MerchantSku:
      type: object
      properties:
        skuId:
          type: integer
        specJson:
          type: object
          additionalProperties:
            type: string
        price:
          type: number
        stock:
          $ref: '#/components/schemas/MerchantStock'
    MerchantStock:
      type: object
      properties:
        available:
          type: integer
          minimum: 0
        locked:
          type: integer
          minimum: 0
    CreateMerchantProductRequest:
      type: object
      required: [categoryId, title, description, mainImage, skus]
      properties:
        categoryId:
          type: integer
        title:
          type: string
        description:
          type: string
        mainImage:
          type: string
        skus:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/CreateMerchantSkuRequest'
    CreateMerchantSkuRequest:
      type: object
      required: [specJson, price, stock]
      properties:
        specJson:
          type: object
          additionalProperties:
            type: string
        price:
          type: number
        stock:
          $ref: '#/components/schemas/CreateMerchantStockRequest'
    CreateMerchantStockRequest:
      type: object
      required: [available]
      properties:
        available:
          type: integer
          minimum: 0
    MerchantProductSubmitAuditResponse:
      type: object
      properties:
        spuId:
          type: integer
        status:
          $ref: '#/components/schemas/ProductStatus'
        message:
          type: string
...
    ProductListResponse:
      type: object
      properties:
        total:
          type: integer
        list:
          type: array
          items:
            $ref: '#/components/schemas/ProductSummary'
    ProductSummary:
      type: object
      properties:
        spuId:
          type: integer
        categoryId:
          type: integer
        categoryName:
          type: string
          nullable: true
        title:
          type: string
        mainImage:
          type: string
        minPrice:
          type: number
          format: double
...
    Address:
      type: object
      properties:
        id:
          type: integer
        receiverName:
          type: string
        phone:
          type: string
        province:
          type: string
        city:
          type: string
        district:
          type: string
        detail:
          type: string
        isDefault:
          type: boolean
    CreateAddressRequest:
      type: object
      required: [receiverName, phone, province, city, district, detail]
      properties:
        receiverName:
          type: string
        phone:
          type: string
        province:
          type: string
        city:
          type: string
        district:
          type: string
        detail:
          type: string
        isDefault:
          type: boolean
    UpdateAddressRequest:
      type: object
      properties:
        receiverName:
          type: string
        phone:
          type: string
        province:
          type: string
        city:
          type: string
        district:
          type: string
        detail:
          type: string
        isDefault:
          type: boolean
    CreateOrderRequest:
      type: object
      required: [addressId, items]
      properties:
        addressId:
          type: integer
        items:
          type: array
          items:
            type: object
            required: [skuId, quantity]
            properties:
              skuId:
                type: integer
              quantity:
                type: integer
        remark:
          type: string
    OrderStatus:
      type: string
      enum:
        - PENDING_PAYMENT
        - PAID
        - PENDING_SHIPMENT
        - SHIPPED
        - COMPLETED
        - CANCELLED
        - REFUNDING
        - REFUNDED
    Order:
      type: object
      properties:
        orderId:
          type: integer
        orderNo:
          type: string
        status:
          $ref: '#/components/schemas/OrderStatus'
        totalAmount:
          type: number
        items:
          type: array
          items:
            type: object
            properties:
              skuId:
                type: integer
              title:
                type: string
              price:
                type: number
              quantity:
                type: integer
        createdAt:
          type: string
          format: date-time